Implement add and delete restaurant for internal user
