--- food-ordering-app/zwiggy/app/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import add_restaurant
-
-urlpatterns = [
-    path('addRestaurant/', add_restaurant)
-]
