+++ food-ordering-app/zwiggy/zwiggy/urls.py
@@ -18,5 +18,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('app/', include('app.urls'))
+    path('internal/', include('internal.urls'))
 ]
